Router get

+++ tests/test_security.py
@@ -16,9 +16,7 @@ def test_jwt():
 
 
 def test_jwt_invalid_token(client):
-    response = client.delete(
-        '/users/1', headers={'Authorization': 'Bearer invalid-token'}
-    )
+    response = client.delete('/users/1', headers={'Authorization': 'Bearer invalid-token'})
 
     assert response.status_code == HTTPStatus.UNAUTHORIZED
     assert response.json() == {'detail': 'Could not validate credentials'}
